LPD-76553 Move types to wizard component
