+++ modules/apps/export-import/export-import-web/src/main/resources/META-INF/resources/revamp/js/components/Footer.tsx
@@ -9,7 +9,7 @@ import ClayLink from '@clayui/link';
 import React from 'react';
 
 import BasePage from '../components/BasePage';
-import {IGenericStepProps} from '../types';
+import {IGenericStepProps} from '../components/Wizard';
 
 const Footer: React.FC<
 	{children?: React.ReactNode | undefined} & IGenericStepProps
